// apps/template-react/frontend/src/gen/models/UserSchema.ts
// @ts-nocheck

/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { UserType } from "./UserType.ts";

/**
 * @description User schema for API responses - NEVER expose password_hash
*/
export type UserSchema = {
    id?: (string | null);
    /**
     * @type string | undefined
    */
    user_type?: UserType;
    /**
     * @default "user"
     * @type string | undefined
    */
    role?: string;
    created_at?: (string | null);
    updated_at?: (string | null);
    display_name?: (string | null);
    username?: (string | null);
    avatar_url?: (string | null);
    language_code?: (string | null);
    timezone?: (string | null);
    birth_date?: (string | null);
    male?: (boolean | null);
    telegram_id?: (number | null);
    tg_first_name?: (string | null);
    tg_last_name?: (string | null);
    tg_username?: (string | null);
    tg_language_code?: (string | null);
    tg_is_premium?: (boolean | null);
    tg_photo_url?: (string | null);
    tg_is_bot?: (boolean | null);
    tg_added_to_attachment_menu?: (boolean | null);
    tg_allows_write_to_pm?: (boolean | null);
    email?: (string | null);
    /**
     * @default false
     * @type boolean | undefined
    */
    email_verified?: boolean;
    email_verified_at?: (string | null);
    last_login_at?: (string | null);
    /**
     * @default 0
     * @type integer | undefined
    */
    current_streak?: number;
    /**
     * @default 0
     * @type integer | undefined
    */
    best_streak?: number;
    /**
     * @default 0
     * @type integer | undefined
    */
    total_active_days?: number;
    last_activity_date?: (string | null);
    is_onboarded?: (boolean | null);
    is_terms_accepted?: (boolean | null);
    referral_code?: (string | null);
};